refactored trader, binance, importer, and simulator into stand-alone command-line applications
