--- univocity-trader-binance-example/pom.xml
@@ -1,26 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>com.univocity</groupId>
-		<artifactId>univocity-trader</artifactId>
-		<version>${revision}</version>
-	</parent>
-
-	<artifactId>univocity-trader-binance-example</artifactId>
-	<version>${revision}</version>
-
-	<name>univocity-trader-binance-example</name>
-	<packaging>jar</packaging>
-
-	<dependencies>
-		<dependency>
-			<groupId>com.univocity</groupId>
-			<artifactId>univocity-trader-binance</artifactId>
-			<version>${revision}</version>
-		</dependency>
-	</dependencies>
-
-</project>
